Remove in-memory loading of the collection
Previously we needed to decrease the position when using offset if the
element has no rank, is not in the current collection (when grouped) or
if it's current position is after the new position. The only time we
actually want to use the position as an offset is if it's in front of the new
position.

Instead of making this distinction we now just remove the element when
querying for the models and always use position - 1. This allows us to
drop the in-memory use of `map` and `index` and the complex
logic from handling all the above-mentioned edge cases.