Allow calling `move_to` with a position out of bounds
If the `move_to` (or `move_to!`) method is called with a position that
exceeds the collection's size, lexorank would have positioned the record
in the middle (using 'U' as the rank).

With this change calling it with such a position will position the
record at the end of the collection.